refactor: code style (#298)
* refactor: code style

* refactor: useProfileStore
